% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{addReadsFeatures}
\alias{addReadsFeatures}
\title{Add reads to the features in the annotation `data.tree`. Also add a separate node with "Unassigned" reads/features}
\usage{
addReadsFeatures(
  tree,
  mappedFeaturesDF,
  featuresCol = "Features",
  readsCol = "Reads"
)
}
\arguments{
\item{tree}{a `data.tree` object}

\item{mappedFeaturesDF}{a `data.frame` with "Reads" and mapped "Features" columns.}

\item{featuresCol}{Column with features}

\item{readsCol}{column with reads}
}
\value{

}
\description{
Add reads to the features in the annotation `data.tree`. Also add a separate node with "Unassigned" reads/features
}
\seealso{
data.tree
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
