% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorList.R
\name{addReadsToTree}
\alias{addReadsToTree}
\title{Add reads to the tree}
\usage{
addReadsToTree(
  fL,
  mappedFeaturesDF,
  featuresCol = "transcript_id",
  readsCol = "seq",
  unassigned = FALSE,
  extraTreeBranch = NULL,
  ...
)
}
\arguments{
\item{fL}{factorList of the annotation}

\item{mappedFeaturesDF}{`DFrame` with Sequences and transcript IDs}

\item{featuresCol}{Feature name column}

\item{readsCol}{Column with sequences}

\item{unassigned}{If unassigned reads to be included in the tree or not}

\item{extraTreeBranch}{Additional sequences as `FactorList` to be added to
the tree. One can obtain unaligned reads using `getReadsFromBam` to obtain
reads from bam files.}

\item{...}{Other parameters to be passed to `fList2tree`}
}
\value{

}
\description{
Add reads to the tree
}
