% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/plotting.R
\name{autoLayout}
\alias{autoLayout}
\alias{autoLayout}
\title{autoLayout}
\usage{
autoLayout(nb, byrow = FALSE)

autoLayout(nb, byrow = FALSE)
}
\arguments{
\item{nb}{Minimum number of frames}

\item{byrow}{Whether to fill frames by row instead of by column (default FALSE, i.e. fill by column)}

\item{nb}{minimum number of panels}
}
\value{
The total number of frames in the layout.

The total number of frames in the layout
}
\description{
Creates a layout for a given (minimum) number of frames

Creates a tiling layout for the given number of panels
}
