% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{base_ratio_proportion_plot}
\alias{base_ratio_proportion_plot}
\title{Make filtering information information data frame from JSON file}
\usage{
base_ratio_proportion_plot(
  df,
  bpPosCol,
  A_col,
  T_col,
  G_col,
  C_col,
  GC_col = NULL,
  N_col,
  isRatio = FALSE
)
}
\arguments{
\item{json}{A JSON file with QC information}
}
\value{
A data frame with FASTQ file information
}
\description{
Make filtering information information data frame from JSON file
}
\examples{
\dontrun{
# Input

# Analysis

# Output
}

}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
