% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep.R
\name{bed2gtf}
\alias{bed2gtf}
\title{bed2gtf}
\usage{
bed2gtf(b, fieldsToInclude = c("transcript_id"))
}
\arguments{
\item{b}{A GRanges object in bed format (as produced by `rtracklayer::import.bed`)}

\item{fieldsToInclude}{The additional columns of `b` to include in the results. By default, only the transcript_id is taken (from the `name` column)}
}
\value{
A GRanges object that can be saved with `rtracklayer::export.gff`
}
\description{
Converts a bed-format GRanges object to a GTF format GRanges object.
}
