% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep.R
\name{buildSrcTree}
\alias{buildSrcTree}
\title{buildSrcTree}
\usage{
buildSrcTree(a, metatypes = NULL, clusters = NULL)
}
\arguments{
\item{a}{An annotation GRanges object}

\item{metatypes}{An optional named vector linking different values of
`transcript_type` in `a` to meta-types. If omitted, will be built
using pre-built grouping scheme.}

\item{clusters}{An optional named vector linking different values of 
`transcript_id` or `gene_id` in `a` to clusters.}
}
\value{
a `data.`
}
\description{
buildSrcTree
}
