% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{calcNormFactors.shortRNAexp}
\alias{calcNormFactors.shortRNAexp}
\title{calcNormFactors.shortRNAexp}
\usage{
calcNormFactors.shortRNAexp(object, method = "TMM",
  normalizeOnType = NULL, normalizeOnStatus = c("unknown", "ambiguous",
  "unique"), ...)
}
\arguments{
\item{object}{An object of class shortRNAexp}

\item{method}{Normalization method (any accepted by edgeR). Defaults to "TMM".}

\item{normalizeOnType}{Types of sequences (e.g. miRNA, tRNA, etc.) to consider for calculating normalization factors (default all).}

\item{normalizeOnStatus}{Sequence statuses to consider for calculating normalization factors (default all mapped)}

\item{...}{Further arguments passed on to edgeR's generic calcNormFactors.}
}
\value{
The shortRNAexp object with updated norm.factors and lib.sizes slots.
}
\description{
Calculates the normalization factors for a shortRNAexp dataset.
}
