% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{capitalizeRead}
\alias{capitalizeRead}
\title{capitalizeRead}
\usage{
capitalizeRead(read, cigar, indels = FALSE)
}
\arguments{
\item{read}{A character vector of length 1 containing the sequnce.}

\item{cigar}{A character vector of length 1 containing the cigar.}

\item{indels}{Logical; whether to also indicate insertions/deletions
(default FALSE). This cannot be enablied if the read already includes
 any hyphen ('-').}
}
\value{
A character vector of length 1 containing the formatted sequence.
}
\description{
Alters a sequence read capitalization according to it's
alignment characteristics (cigar).
}
