% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{checkSizeSelection}
\alias{checkSizeSelection}
\title{checkSizeSelection}
\usage{
checkSizeSelection(o, type = NULL, status = c("unknown", "ambiguous",
  "unique"), trim = TRUE, onlyDetectedInAll = TRUE)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{type}{Types of sequences to be fetched (e.g. miRNA, tRNA, etc.), fetches all by default.}

\item{status}{Character vector of seq statuses to be selected (any combination of "unmapped","unknown","ambiguous", and "unique"). Defaults to all mapped sequences.}

\item{trim}{Logical; whether to trim the 5\% most extreme values on each side before calculating mean (default TRUE).}
}
\description{
Plots different measures of abundance according to sequence length across samples. This essentially calls `plotSizeAbundance` several times on different aggregations.
}
