% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{collapsed2countMatrix}
\alias{collapsed2countMatrix}
\title{collapsed2countMatrix}
\usage{
collapsed2countMatrix(seqcounts.files, output.file, shell = "bash")
}
\arguments{
\item{seqcounts.files}{A character vector of path to .seqcounts files (i.e. files produced by `collapseFastq`).}

\item{output.file}{Path to the file where the output matrix will be saved.}

\item{shell}{Shell executable, default 'bash'.}
}
\description{
Creates a count matrix from a list of .seqcounts files
}
