% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{collapseFastq}
\alias{collapseFastq}
\title{collapseFastq}
\usage{
collapseFastq(files, minLength = 15, maxLength = 49, shell = "bash")
}
\arguments{
\item{files}{A character vector of path to fastq (or fastq.gz) files.}

\item{minLength}{The minimum length for a read to be considered (default 15).}

\item{maxLength}{The maximum length for a read to be considered (default 49).}

\item{shell}{Shell executable, default 'bash'.}
}
\description{
Extract unique reads and their counts from (adapter-trimmed) fastq files, and save the contents in .seqcounts files.
}
