% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{combineFeaturesDF}
\alias{combineFeaturesDF}
\title{Combine a list of `data.frame` into one `data.tree`. `data.frame` should have a column "pathString", or specify.}
\usage{
combineFeaturesDF(dfList, rootName = "root", ...)
}
\arguments{
\item{dfList}{a list of data frames}

\item{rootName}{name for the root}
}
\value{

}
\description{
Combine a list of `data.frame` into one `data.tree`. `data.frame` should have a column "pathString", or specify.
}
\examples{
mm10 <- combineFeaturesDF(dfList = list(tRNA$df, miRNA$df), rootName = "mm10")
}
\seealso{
data.tree plyr
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
