% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep2.R
\name{convertRSStoDSS}
\alias{convertRSStoDSS}
\title{Convert a RNAStringSet to DNAStringSet}
\usage{
convertRSStoDSS(rss)
}
\arguments{
\item{rss}{RNAStringSet}
}
\value{

}
\description{
Convert a RNAStringSet to DNAStringSet
}
