% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{defaultAssignRules}
\alias{defaultAssignRules}
\title{defaultAssignRules}
\usage{
defaultAssignRules(rules = list())
}
\arguments{
\item{rules}{An optional named list of rules, overwriting parts of the
default rules.}
}
\value{
A list
}
\description{
defaultAssignRules
}
\examples{
rules <- defaultAssignRules()
}
