% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{dupPlotDF}
\alias{dupPlotDF}
\title{Reads duplication data frame from JSON file}
\usage{
dupPlotDF(json)
}
\arguments{
\item{json}{A JSON file with QC information from \code{\link{qcFastq}}}
}
\value{
An interactive plot

#' @examples
# Input
fq1 <- system.file("extdata", "Fox3_Std_small.fq.gz", package = "Rfastp")
outDir <- tempdir()

# Analysis
qc_res <- qcFastq(file = fq1, outDir = outDir)
df <- dupPlotDF(qc_res$json)

# Output
df
}
\description{
Reads duplication data frame from JSON file
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
