% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallRNAexp.R
\name{estimateComposition}
\alias{estimateComposition}
\title{estimateComposition}
\usage{
estimateComposition(o, m = NULL, abridgedTypes = NULL)
}
\arguments{
\item{o}{An object of class smallRNAexp}

\item{m}{An optional seqcount table. If ommited, o@seqcounts will be used.}

\item{abridgedTypes}{An optional named vector of RNA types to be aggregated.}
}
\value{
A list containing two data.frames.
}
\description{
Estimates the relative abundance of different types of RNAs populating a dataset.
}
