% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{fastaToTree}
\alias{fastaToTree}
\title{Convert the annotation fasta file into tree obtained from `prepareAnnotation` function}
\usage{
fastaToTree(file, root)
}
\arguments{
\item{root}{root name: longRNA or miRNA or tRNA}

\item{fasta}{a link to the fasta file}
}
\value{

}
\description{
Convert the annotation fasta file into tree obtained from `prepareAnnotation` function
}
\seealso{
Biostrings data.tree plyr
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
