% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqCounts.R
\name{fastq2SeqCountMatrix}
\alias{fastq2SeqCountMatrix}
\title{fastq2SeqCountMatrix}
\usage{
fastq2SeqCountMatrix(
  files,
  minLength = 15L,
  maxLength = 200L,
  discardBelowCount = 2L,
  BPPARAM = BiocParallel::SerialParam(progress = TRUE),
  format = NULL,
  asSparse = FALSE
)
}
\arguments{
\item{files}{A character vector containing the path to each fastq file.}

\item{minLength}{The minimum length for a sequence to be included (def. 15)}

\item{maxLength}{The maximum length for a sequence to be included (default
Inf, i.e. no maximum)}

\item{discardBelowCount}{Discard sequences that appear less than this number
of times across all samples. The default value is 2, which means that
sequences that appear only once are discarded.}

\item{BPPARAM}{BiocParallel BPPARAM object for multithreading (default no
multithreading). Note that multithreading will require more memory and is
therefore not recommended for larger datasets.}

\item{format}{File format, either 'fasta' or 'fastq'. If omitted will be
automatically detected from the first file's name.}

\item{asSparse}{Logical; whether to convert the matrix to sparse format (not
yet entirely supported by downstream functions)}
}
\value{
a matrix, with unique sequences as rows and samples (i.e. file
  basenames) as column.
}
\description{
Collapses a set of `.fastq` (or `fastq.gz`) files to a matrix of counts per
 unique sequence.
}
