% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeFL.R
\name{featureToFL}
\alias{featureToFL}
\title{Convert the annotation fasta file into tree obtained from `prepareAnnotation` function}
\usage{
featureToFL(anno, root)
}
\arguments{
\item{root}{root name: longRNA or miRNA or tRNA}

\item{fasta}{a link to the fasta file}
}
\value{

}
\description{
Convert the annotation fasta file into tree obtained from `prepareAnnotation` function
}
\examples{
tRNA <- fastaToTree(file = "test/tRNAs.modified.fa", root = "tRNA")
miRNA.mirBase <- fastaToTree(file = "test/miRNAs.modified.fa", root = "miRNA")
rRNA <- fastaToTree(file = "test/Mus_musculus/rRNAdb/", root = "rRNA")
longRNA.gencode <- fastaToTree(file = "test/gencode.vM22.chr_patch_hapl_scaff.annotation.gff3.gz", root = "longRNA")
}
\seealso{
Biostrings data.tree plyr
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
