% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlap.R
\name{findoverlaps.bam.featureAnnotation}
\alias{findoverlaps.bam.featureAnnotation}
\title{findoverlaps.bam.featureAnnotation}
\usage{
findoverlaps.bam.featureAnnotation(bamFile, featureAnnotation,
  overlapBy = 0.5)
}
\arguments{
\item{bamFile}{A character vector of path to BAM file.}

\item{featureAnnotation}{A GRanges object as produced by prepareAnnotation function.}

\item{overlapBy}{Minimum proportion of the read inside the overlapping feature, ranging between 0 and 1 (default: 0.5).}
}
\value{
A dataframe.
}
\description{
Map a BAM file to the features file
}
