% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_plg.R
\name{fList2tree}
\alias{fList2tree}
\title{fList2tree}
\usage{
fList2tree(fL, addRoot = TRUE, collapseSingles = FALSE, root = "ROOT")
}
\arguments{
\item{fL}{A `FactorList` object}

\item{addRoot}{Logical; whether to add a root node}

\item{collapseSingles}{Logical; wether to remove intermediate nodes of single
childs}
}
\value{
A `phylo` object
}
\description{
Converts a `FactorList` to a `phylo` object.
}
