% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortRNAexp.R
\name{getAggCounts}
\alias{getAggCounts}
\title{getAggCounts}
\usage{
getAggCounts(o, type = NULL, ambiguous = is.null(type),
  normalized = FALSE)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{type}{Types of features to be fetched (e.g. miRNA, tRNA, etc.), fetches all by default.}

\item{ambiguous}{Logical; whether to include ambiguous features (default TRUE if `type` is NULL).}

\item{normalized}{Logical; whether to return the normalized counts (default FALSE)}
}
\value{
A count matrix.
}
\description{
Returns the read counts of aggregated features
}
