% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{getAggOffset}
\alias{getAggOffset}
\title{getAggOffset}
\usage{
getAggOffset(o, offset)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{offset}{The offset matrix to aggregate}
}
\value{
A numeric matrix.
}
\description{
Aggregates offsets for cqn normalization using weighted means
}
