% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallRNAexp.R
\name{getComposition}
\alias{getComposition}
\title{getComposition}
\usage{
getComposition(o, abridged = FALSE, exclude = c("unknown", "ambiguous"),
  normalized = FALSE, scale = TRUE)
}
\arguments{
\item{o}{An object of class smallRNAexp}

\item{abridged}{Logical; whether to return the abridged RNA types (default FALSE)}

\item{exclude}{Character vector of RNA classes to exclude (by default, excludes only ambiguous and unknown).}

\item{normalized}{Logical; whether to use normalized counts rather than raw counts to calculate abundances (default FALSE)}

\item{scale}{Logical; whether to linearly scale the abundances so that they sum to 1 (default FALSE)}
}
\value{
A matrix of relative abundances.
}
\description{
Get the relative abundance of different types of RNAs populating a smallRNAexp dataset.
}
