% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortRNAexp.R
\name{getCompositionComponent}
\alias{getCompositionComponent}
\title{getCompositionComponent}
\usage{
getCompositionComponent(o, summarytype = "MDS", ...)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{summarytype}{The type of dimensionality reduction to use. Either "pca" (reporting the first two principal components), "mds" (using a single variable), or "polar" polar coordinates of the first 2 principal components. Default is "mds".}

\item{...}{Any argument passed to the `getComposition` function.}
}
\value{
A matrix with the composition varible(s).
}
\description{
Retrieves component variables explaining the variation in RNA type composition.
}
