% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep2.R
\name{getDB}
\alias{getDB}
\title{Obtain databases for species (currently supported for mouse only)}
\usage{
getDB(
  species = "mmu",
  genomeVersion = "GRCm38",
  tRNA_addCCA = TRUE,
  tRNA_includeMt = TRUE,
  rRNA_release = "138.1"
)
}
\arguments{
\item{species}{GRCm38 or GRCm39}
}
\value{
A list of databases (DNAstringSet or GRanges)
}
\description{
Obtain databases for species (currently supported for mouse only)
}
