% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortRNAexp.R
\name{getFeatureCounts}
\alias{getFeatureCounts}
\title{getFeatureCounts}
\usage{
getFeatureCounts(o, feature, exact = TRUE, normalized = FALSE,
  allowAmbiguous = FALSE)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{feature}{The name of the feature (e.g. gene name) or, if !`exact`), a regular expression to be evaluated against all feature names.}

\item{exact}{Logical; indicates whether `feature` should be interpreted as an exact name (default) rather than a regular expression.}

\item{normalized}{Logical; whether to return the normalized counts (default FALSE)}

\item{allowAmbiguous}{Logical; indicates whether to allow selection of ambiguous sequences when using exact=FALSE (default FALSE).}
}
\value{
A count matrix.
}
\description{
Returns the read counts of sequences associated to a feature
}
