% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallRNAexp.R
\name{getFeatureSeqs}
\alias{getFeatureSeqs}
\title{getFeatureSeqs}
\usage{
getFeatureSeqs(o, feature, exact = TRUE, allowAmbiguous = FALSE,
  ignore.case = TRUE)
}
\arguments{
\item{o}{An object of class smallRNAexp}

\item{feature}{The name of the feature (e.g. gene name) or, if !`exact`), a regular expression to be evaluated against all feature names.}

\item{exact}{Logical; indicates whether `feature` should be interpreted as an exact name (default) rather than a regular expression.}

\item{allowAmbiguous}{Logical; indicates whether to allow selection of ambiguous sequences when using exact=FALSE (default FALSE).}

\item{ignore.case}{Logical; indicates whether to ignore case when using exact=FALSE (default TRUE)}
}
\value{
A vector of sequences
}
\description{
Returns the sequences associated to a feature
}
