% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep2.R
\name{getmiRNA}
\alias{getmiRNA}
\title{Obtain miRNA GRanges from mirBase}
\usage{
getmiRNA(sp = "mmu")
}
\arguments{
\item{sp}{3 alphabet species code. Please check `data(species)`}
}
\value{
a list of GRanges annotation of miRNAs, miRNAs DNAStringSet,
and miRNA hairpins miRNAs DNAStringSet
}
\description{
Obtain miRNA GRanges from mirBase
}
\examples{
mir_mouse <- getmiRNA(sp = "mmu")
mir_human <- getmiRNA(sp = "hsa")
}
