% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep2.R
\name{getMttRNA}
\alias{getMttRNA}
\title{Obtain mitiochondrial tRNAs from mitotRNA batabase}
\usage{
getMttRNA(sp = "Mus musculus", addCCA = TRUE)
}
\arguments{
\item{sp}{Scientific name of species}

\item{addCCA}{Whether to add CCA modifications to sequences or not.
Default: TRUE}
}
\description{
Obtain mitiochondrial tRNAs from mitotRNA batabase
}
