% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{getOverlapValidity}
\alias{getOverlapValidity}
\title{getOverlapValidity}
\usage{
getOverlapValidity(sources, rules = defaultAssignRules())
}
\arguments{
\item{sources}{the output of overlapWithTx2}

\item{rules}{}
}
\value{
Data table (same format as `sources`)
}
\description{
getOverlapValidity
}
