% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getQualitativePalette}
\alias{getQualitativePalette}
\title{getQualitativePalette}
\usage{
getQualitativePalette(nbcolors)
}
\arguments{
\item{nbcolors}{number of colors (from 1 to 22)}
}
\value{
A vector of colors
}
\description{
Returns a qualitative color palette of given size
}
