% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorList.R
\name{getReadsFromBam}
\alias{getReadsFromBam}
\title{Get reads from BAM file for a specific samFlag}
\usage{
getReadsFromBam(bam, flag = 4, label = "unaligned")
}
\arguments{
\item{bam}{A label to be given to reads}

\item{flag}{Flag for which reads are to be obtained. Default: 4 (Unaligned)}
}
\value{
A `FactorList` of reads with labels
}
\description{
Get reads from BAM file for a specific samFlag
}
