% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep2.R
\name{getrRNA}
\alias{getrRNA}
\title{Obtain rRNAs from SILVA database}
\usage{
getrRNA(sp = "Mus musculus", release = "138.1")
}
\arguments{
\item{sp}{Scientific names of the species}

\item{release}{Which release should be used from 128, 132 and 138.1}
}
\value{
DNAStringSet of rRNAs
}
\description{
Obtain rRNAs from SILVA database
}
