% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortRNAexp.R
\name{getSeqsByType}
\alias{getSeqsByType}
\title{getSeqsByType}
\usage{
getSeqsByType(o, type = NULL, status = NULL)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{type}{Types of sequences to be fetched (e.g. miRNA, tRNA, etc.), fetches all by default.}

\item{status}{Character vector of seq statuses to be selected (any combination of 'unmapped','invalid','unknown','ambiguous', and 'unique'). Fetches all by default.}
}
\value{
A character vector of sequences.
}
\description{
Returns sequences matching filters
}
