% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsDKT.R
\name{indexRsubread}
\alias{indexRsubread}
\title{Build index for Rsubread}
\usage{
indexRsubread(
  fastaGenome = NULL,
  refGenome = "mm10",
  gencodeRelease = "M18",
  outPath = getwd(),
  basename = "refRsubread",
  ...
)
}
\arguments{
\item{fastaGenome}{path to a gzipped fasta file (optional if want to build genome other than mm9, mm10, hg19 or hg38).}

\item{refGenome}{mm9/mm10/hg19/hg38. `fasta` file is downloaded from gencode (https://gencodegenes.org)}

\item{gencodeRelease}{Release of the genome to be downloaded.}

\item{outPath}{Output direcory where genome is downloaded and index is build.}

\item{basename}{Basename for the Rsubread index.}

\item{...}{Other options for `Rsubread::buildindex()`.}
}
\value{

}
\description{
Build index for Rsubread
}
\seealso{
Rsubread
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
