% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{isPrimaryPiRNA}
\alias{isPrimaryPiRNA}
\title{isPrimaryPiRNA}
\usage{
isPrimaryPiRNA(src, allowRevComp = FALSE, length = 26:32)
}
\arguments{
\item{src}{A data.frame of overlaps, as produced by
\code{\link{overlapWithTx}}}

\item{allowRevComp}{Logical; whether to allow the condition to be fulfilled
on the reverse complement sequences}

\item{length}{length range}

\item{rules}{Assignment rules (see \code{\link{defaultAssignmentRules}})}
}
\value{
A logical vector of the same length as there are rows in `src`
}
\description{
isPrimaryPiRNA
}
