% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{isValidMiRNA}
\alias{isValidMiRNA}
\title{isValidMiRNA}
\usage{
isValidMiRNA(src, length = 19:24, minOverlap = 16L, maxNonOverlap = 3L)
}
\arguments{
\item{src}{A data.frame of overlaps with miRNAs, as produced by
\code{\link{overlapWithTx}}}

\item{length}{length range}

\item{minOverlap}{minimum overlap with the annotated mature miRNA}

\item{maxNonOverlap}{maximum number of nucleotides beyond the feature}

\item{rules}{Assignment rules (see \code{\link{defaultAssignmentRules}})}
}
\value{
A logical vector of the same length as there are rows in `src`
}
\description{
isValidMiRNA
}
