% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{isValidOverlap}
\alias{isValidOverlap}
\title{isValidOverlap}
\usage{
isValidOverlap(srcs, rules = defaultAssignRules)
}
\arguments{
\item{srcs}{A data.frame of overlaps, as produced by
\code{\link{overlapWithTx}}}

\item{rules}{Assignment rules (see \code{\link{defaultAssignmentRules}})}
}
\value{
A logical vector of the same length as there are rows in `srcs`
}
\description{
isValidOverlap
}
