% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{listFilesFTP}
\alias{listFilesFTP}
\title{List files on a FTP server}
\usage{
listFilesFTP(url)
}
\arguments{
\item{url}{A url of FTP location}
}
\value{
A list of files
}
\description{
List files on a FTP server
}
\examples{
# Input
url <- "ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M18/"

# Analysis
files <- listFilesFTP(url)

# Output
files
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
