% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_plg.R
\name{longestCommonString}
\alias{longestCommonString}
\title{longestCommonString}
\usage{
longestCommonString(x, delim = "")
}
\arguments{
\item{x}{A character vector}

\item{delim}{The delimiter (by default, splits all letters)}
}
\value{
A character vector of length 1
}
\description{
Finds the longest common part of a set of character strings
}
\examples{
a <- c("B1/B2/B3", "B1/B2/B3/B4", "B1/B2/B5")
longestCommonString(a, "/")
}
