% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorList.R
\name{longestOrderedOverlap}
\alias{longestOrderedOverlap}
\title{longestOrderedOverlap}
\usage{
longestOrderedOverlap(x)
}
\arguments{
\item{x}{An AtomicList}
}
\value{
An atomic vector
}
\description{
Returns the longest ordered overlap between elements of an atomic list
}
\examples{
il <- IntegerList(list(1:4, 1:5, c(1:3, 6)))
longestOrderedOverlap(il)
}
