% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{make_before_filt_df}
\alias{make_before_filt_df}
\title{Make before QC information data frame from JSON file}
\usage{
make_before_filt_df(json)
}
\arguments{
\item{json}{A JSON file with QC information}
}
\value{
A data frame with FASTQ file information
}
\description{
Make before QC information data frame from JSON file
}
\examples{
\dontrun{
# Input

# Analysis

# Output

}

}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
