% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{make_dup_plot}
\alias{make_dup_plot}
\title{Make duplicated reads plot from JSON file}
\usage{
make_dup_plot(json)
}
\arguments{
\item{json}{A JSON file}
}
\value{
An interactive plot

#' @examples
# Input
fq1 <- system.file("extdata", "Fox3_Std_small.fq.gz", package = "Rfastp")
outDir <- tempdir()

# Analysis
qc_res <- qc_fastq(file = fq1, outDir = outDir)
p <- make_dup_plot(json = qc_res$json)

# Output
p
}
\description{
Make duplicated reads plot from JSON file
}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
