% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{make_general_df}
\alias{make_general_df}
\title{Make general information data frame from JSON file}
\usage{
make_general_df(json)
}
\arguments{
\item{json}{A JSON file with QC information}
}
\value{
A data frame with FASTQ file information
}
\description{
Make general information data frame from JSON file
}
\examples{
\dontrun{
# Input
fq1 <- system.file("extdata", "Fox3_Std_small.fq.gz", package = "Rfastp")

# Analysis
qc_res <- qc_fastq(file = fq1)
df <- make_general_df(json = qc_res$json)

# Output
df
}

}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
