% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{make_read_length_plot}
\alias{make_read_length_plot}
\title{Read length distribution plot}
\usage{
make_read_length_plot(before_trim, after_trim)
}
\arguments{
\item{before_trim}{A dataframe with 2 columns: read_length and num_reads}

\item{after_trim}{A dataframe with 2 columns: read_length and num_reads}
}
\value{
An interactive barplot
}
\description{
Read length distribution plot
}
\examples{
\dontrun{
# Input

# Analysis

# Output
}

}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
