% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{make_read_qual_plot}
\alias{make_read_qual_plot}
\title{Make filtering information information data frame from JSON file}
\usage{
make_read_qual_plot(json, which = "before")
}
\arguments{
\item{json}{A JSON file with QC information}
}
\value{
A data frame with FASTQ file information
}
\description{
Make filtering information information data frame from JSON file
}
\examples{
\dontrun{
# Input

# Analysis

# Output

}

}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
