% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{makeGeneralDF}
\alias{makeGeneralDF}
\title{Make general information data frame from JSON file}
\usage{
makeGeneralDF(json)
}
\arguments{
\item{json}{A JSON file with QC information from \code{\link{qcFastq}}}
}
\value{
A data frame with FASTQ file information
}
\description{
Make general information data frame from JSON file
}
\examples{
# Input
fq1 <- system.file("extdata", "Fox3_Std_small.fq.gz", package = "Rfastp")
outDir <- tempdir()

# Analysis
qc_res <- qcFastq(file = fq1, outDir = outDir)
df <- makeGeneralDF(json = qc_res$json)

# Output
df
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
