% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{makeTracks}
\alias{makeTracks}
\title{Make tracks with reads}
\usage{
makeTracks(
  gr,
  counts,
  features,
  bamFile = NULL,
  param = ScanBamParam(what = c("cigar", "pos", "seq")),
  plotCoverage = TRUE
)
}
\arguments{
\item{gr}{`GRanges` object with the genomic coordinates to be plotted}

\item{counts}{a `dataframe` of samples by features counts}

\item{features}{`GRanges` of features}

\item{bamFile}{path to a BAM file}

\item{param}{parameters for bam file}

\item{plotCoverage}{Logical. Wether to plot coverage or not.}
}
\value{
plot with tracks
}
\description{
Make tracks with reads
}
