% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{maketrans}
\alias{maketrans}
\title{maketrans}
\usage{
maketrans(tcol, alpha = 100)
}
\arguments{
\item{tcol}{the color to be made transparent, specified by any R color specification (see \code{\link[grDevices]{col2rgb}}}

\item{alpha}{an integer from 0 to 255 indicating the degree of transparency (0 is totally transparent, 255 is totally opaque). Default 100.}
}
\value{
the transparent color code
}
\description{
Makes a given color transparent.
}
\examples{
maketrans("black")

}
