% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_plg.R
\name{mergeAtomicLists}
\alias{mergeAtomicLists}
\title{mergeAtomicLists}
\usage{
mergeAtomicLists(x, y)
}
\arguments{
\item{x}{An `AtomicList` object}

\item{y}{An `AtomicList` object}
}
\value{
An `AtomicList` object of the same length as `x` and `y`
}
\description{
Pairwise merging of two atomic lists of the same length. Does not perform 
any matching of the names, but triggers an error if names are present and 
not in the same order.
}
\examples{
f1 <- FactorList(list(LETTERS[1:3],LETTERS[1:4]))
f2 <- FactorList(list(LETTERS[4:6],LETTERS[7:9]))
mergeAtomicLists(f1,f2)
}
