% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortRNAexp.R
\name{mergeshortRNADatasets}
\alias{mergeshortRNADatasets}
\title{mergeshortRNADatasets}
\usage{
mergeshortRNADatasets(o1, o2, by = "feature", ...)
}
\arguments{
\item{o1}{An object of class shortRNAexp.}

\item{o2}{An object of class shortRNAexp.}

\item{by}{Level at which to aggregate; either 'feature' (default) or 'sequence'.}

\item{...}{Arguments passed to the merge function calls. This can be used to pass for instance `all.x` and `all.y` arguments, although the result object is likely not the be compatible with all the packages' functions afterwards.}
}
\value{
An object of class shortRNAexp.
}
\description{
Merges two shortRNAexp objects.
}
