% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep2.R
\name{miRNAcluster}
\alias{miRNAcluster}
\title{Cluster miRNAs located close to each other}
\usage{
miRNAcluster(gr, minGap = 10000)
}
\arguments{
\item{gr}{GRanges of miRNAs}

\item{minGap}{minimum gap between the miRNAs for grouping}
}
\value{
A GRanges object
}
\description{
Cluster miRNAs located close to each other
}
