% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeFL.R
\name{miRNAtoFL}
\alias{miRNAtoFL}
\title{Convert miRNAs annotation GRanges to a Factor List}
\usage{
miRNAtoFL(gr, spMore = TRUE, sp = "mm10", cluster = TRUE)
}
\arguments{
\item{gr}{annotation GRanges with columns: tx_id, tx_biotype, and symbol}

\item{sp}{name of the species}
}
\value{

}
\description{
Convert miRNAs annotation GRanges to a Factor List
}
