% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{msaWrapper}
\alias{msaWrapper}
\title{msaWrapper}
\usage{
msaWrapper(seqs)
}
\arguments{
\item{seqs}{A character vector containing DNA sequences}
}
\value{
A character vector of length=length(seqs), with each sequence offsetted so that they align
}
\description{
Returns multiple alignment of the input sequences
}
