% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsDKT.R, R/misc.R
\name{mylapply}
\alias{mylapply}
\title{Run `lapply` function in parallel, if possible.}
\usage{
mylapply(...)

mylapply(...)
}
\arguments{
\item{any}{parameters for `lapply` or `mclapply`}
}
\value{
A `list`.

A `list`.
}
\description{
Run `lapply` function in parallel, if possible.

Run `lapply` function in parallel, if possible.
}
\author{
Deepak Tanwar (tanward@ethz.ch)

Deepak Tanwar (tanward@ethz.ch)
}
