% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{normalizeCounts}
\usage{
normalizeCounts(counts, normParams)
}
\arguments{
\item{counts}{A numeric count matrix (with samples as columns)}

\item{normParams}{A list of normalization parameters (e.g. the slot `norm` of a `smallRNAexp` object).}
}
\value{
A normalized matrix.
}
\description{
Normlalizes a count matrix using the given norm.factors
}
