% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeFL.R
\name{otherRNA2FL}
\alias{otherRNA2FL}
\title{Convert RNAs annotation GRanges to a Factor List}
\usage{
otherRNA2FL(gr, sp = "mm10")
}
\arguments{
\item{gr}{annotation GRanges with columns: tx_id, tx_biotype, and symbol}

\item{sp}{name of the species}
}
\value{

}
\description{
Convert RNAs annotation GRanges to a Factor List
}
