% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{plotAdapterResults}
\alias{plotAdapterResults}
\title{plotAdapterResults}
\usage{
plotAdapterResults(
  o,
  showCoocurrence = FALSE,
  minFreq = 0.05,
  addSeq = TRUE,
  row_names_width = ifelse(showCoocurrence, 10, 16)
)
}
\arguments{
\item{o}{An object of class `adapterResults`, as produced by `tryAdapters`}

\item{showCoocurrence}{Logical; whether to plot adapter co-occurence instead
of adapter frequency by position (default FALSE)}

\item{minFreq}{Minimum frequency for an adapter to be displayed (default
0.05). Frequency is calculated at the 4th nucleotide from the end of
 the read.}

\item{addSeq}{Logical; whether to add the adapter sequence to the row.names.}

\item{row_names_width}{Width of the row names.}
}
\description{
plotAdapterResults
}
