% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotComposition}
\alias{plotComposition}
\title{plotComposition}
\usage{
plotComposition(o, exclude = NULL, include = NULL,
  normalized = FALSE, abridged = TRUE, scale = TRUE, ...)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{exclude}{Character vector of RNA classes to exclude (by default, excludes only ambiguous and unknown). Trumps `include`.}

\item{include}{Character vector of RNA classes to include (by default, all except those specified in `exclude`).}

\item{normalized}{Logical; whether to use normalized abundances (default FALSE).}

\item{abridged}{Logical; whether to return the abridged RNA types (default TRUE)}

\item{scale}{Logical; whether to linearly scale the abundances so that they sum to 1 (default TRUE)}

\item{...}{any further arguments passed to the plotting functions.}
}
\value{
Nothing, but plots multiple frames.
}
\description{
Plots the relative abundance of different types of RNAs populating a shortRNAexp dataset.
}
