% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotFeature}
\alias{plotFeature}
\title{plotFeature}
\usage{
plotFeature(o, feature = NULL, sequences = NULL, exact = TRUE,
  normalized = TRUE, aggregated = FALSE, plotType = "barplot",
  proportion = FALSE, plotLegend = TRUE, dolog = plotType != "barplot",
  main = feature, ylab = NULL, allowAmbiguous = FALSE,
  alignSeqs = !aggregated, barplotLayout = matrix(c(1, 2, 2), nrow = 1),
  ...)
}
\arguments{
\item{o}{An object of class smallRNAexp}

\item{feature}{The name of the feature (e.g. gene name) or, if !`exact`), a regular expression to be evaluated against all feature names. One and only one of `sequences` and `feature` should be given.}

\item{sequences}{A character vector of the sequences to plot. One and only one of `sequences` and `feature` should be given.}

\item{exact}{Logical; indicates whether `feature` should be interpreted as an exact name (default) rather than a regular expression.}

\item{normalized}{Logical; whether to return the normalized counts (default TRUE)}

\item{aggregated}{Logical; whether to return aggregated counts rather than per-sequence counts (default FALSE)}

\item{plotType}{Either "barplot" (default), "heatmap", or "zheatmap" (heatmap of feature-wise z-scores").}

\item{proportion}{Logical; whether to plot proportions (i.e. sequences summing to 1) rather than actual/normalized counts (default FALSE).}

\item{plotLegend}{Logical; whether to plot the legend (default TRUE); will be plotted on a distinct frame.}

\item{dolog}{Logical; whether to plot log-transformed counts (default FALSE if plotType="barplot", TRUE otherwise).}

\item{main}{Plot title, defaults to the name of the feature queried.}

\item{allowAmbiguous}{Logical; indicates whether to allow selection of ambiguous sequences when using exact=FALSE (default FALSE).}

\item{alignSeqs}{Logical; whether to show sequences as a multiple alignment (default TRUE).}

\item{barplotLayout}{a matrix populated by 1s and 2s indicating the layout for the barplot (2) and legend (1). Defaults to matrix(c(1,2,2),nrow=1). Ignored if plotType!="barplot".}

\item{...}{arguments passed to the plotting function.}
}
\value{
Nothings but generates plots.
}
\description{
Plots the read counts of sequences associated to a given feature (or potentially feature(s) if regular expressions are used), or the aggregated counts if `aggregated`=TRUE.
}
