% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotFeatureCoverage}
\alias{plotFeatureCoverage}
\title{plotFeatureCoverage}
\usage{
plotFeatureCoverage(
  o,
  feature,
  exact = TRUE,
  includeAmbiguous = TRUE,
  logCounts = FALSE,
  main = ""
)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{feature}{The name of the feature}

\item{exact}{Logical; whether the feature name should be interpretated as exact (default TRUE), otherwise as a regular expression.}

\item{includeAmbiguous}{Logical; whether to include ambiguous sequences (default TRUE; plotted a different color)}

\item{logCounts}{Logical; whether to log-transform read counts for plotting (defualt FALSE) color)}
}
\description{
Plots the coverage across a transcript/feature
}
