% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGC}
\alias{plotGC}
\title{plotGC}
\usage{
plotGC(o, aggregateSamples = FALSE, ...)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{aggregateSamples}{Logical; whether to aggregated samples instead of plotting them individually (default).}

\item{...}{Any filtering argument passed to the getSeqCounts function.}
}
\description{
Plots the expression across different GC proportions.
}
