% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMAs}
\alias{plotMAs}
\title{plotMAs}
\usage{
plotMAs(o, normalized = TRUE, type = NULL, status = c("unknown",
  "ambiguous", "unique"), absfc = FALSE, samples = NULL, average = FALSE,
  ...)
}
\arguments{
\item{o}{An object of class smallRNAexp}

\item{normalized}{Logical; whether normalized counts should be used (default TRUE)}

\item{type}{Types of sequences to be fetched (e.g. miRNA, tRNA, etc.), fetches all by default.}

\item{status}{Character vector of seq statuses to be selected (any combination of "unmapped","unknown","ambiguous", and "unique"). Fetches all mapped by default.}

\item{absfc}{Logical; whether to plot absolute log2-foldchange instead of log2-foldchange (default FALSE)}

\item{samples}{Optional numeric vector indicating which sample(s) to compare. If NULL (default), all samples are used.}

\item{...}{any further arguments passed to the scatter.smooth funciton.}
}
\value{
Nothing, but plots multiple frames.
}
\description{
Plots MAs relative to the mean
}
