% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSizeAbundance}
\alias{plotSizeAbundance}
\title{plotSizeAbundance}
\usage{
plotSizeAbundance(o, type = NULL, status = c("unknown", "ambiguous",
  "unique"), normalized = TRUE, plotLengthFrequencies = TRUE,
  onlyDetectedInAll = FALSE, dolog10 = FALSE, agFun = sum,
  plotLegend = NULL, ...)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{type}{Types of sequences to be fetched (e.g. miRNA, tRNA, etc.), fetches all by default.}

\item{status}{Character vector of seq statuses to be selected (any combination of "unmapped","unknown","ambiguous", and "unique"). Defaults to all mapped sequences.}

\item{normalized}{Logical; whether to return the normalized counts (default FALSE)}

\item{plotLengthFrequencies}{Logical; whether to plot the number of sequences with each length (default TRUE)}

\item{onlyDetectedInAll}{Logical; whether use counts only from sequences detected in all samples (default FALSE)}

\item{dolog10}{Logical; whether to use log10-counts (default FALSE).}

\item{...}{arguments passed to the plot function.}
}
\value{
Nothing, but produces a plot.
}
\description{
Plots the abundance according to sequence length across samples
}
