% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep.R
\name{prepareAnnotation}
\alias{prepareAnnotation}
\title{prepareAnnotation}
\usage{
prepareAnnotation(filelist)
}
\arguments{
\item{filelist}{A list of named feature files, as produced by 'getAnnotationFiles', indicating the necessary files. The following slots are required: 'features.gtf','features.gtf','mirbase.gtf','gtRNAdb.fa','gtRNAdb.bed','tRNA.gtf'; the 'pi_precursors.gtf' is optional.}
}
\value{
A GRanges object, and produces modified fasta files.
}
\description{
Prepares an annotation GRanges and modified miRNA/tRNA sequences for a given organism. See `getAnnotationFiles` for obtaining the necessary files.
}
