% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qcFastq}
\alias{qcFastq}
\title{QC of fastq files}
\usage{
qcFastq(
  file,
  outDir = ".",
  name = gsub(pattern = "\\\\.fq|\\\\.fq.gz|\\\\.fastq|\\\\.fastq.gz", replacement = "",
    x = basename(file)),
  adapterSeq = "auto",
  adaptersFa = "",
  minLength = 15,
  nThread = parallel::detectCores(),
  ...
)
}
\arguments{
\item{file}{character vector of file names}

\item{outDir}{output directory. Default: current working directory}

\item{name}{character vector of output file names.}

\item{adapterSeq}{Adapter sequence for read1}

\item{adaptersFa}{Adapter sequence fasta file}

\item{...}{other parameters specific to \code{\link[Rfastp:rfastp]}}
}
\value{
A list of results and location to the JSON results file
}
\description{
QC of fastq files
}
\examples{
# Input
fq1 <- system.file("extdata", "Fox3_Std_small.fq.gz", package = "Rfastp")
outDir <- tempdir()

# Analysis
qc_res <- qcFastq(file = fq1, outDir = outDir)

# Output
qc_res
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
