% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{read_length_hist}
\alias{read_length_hist}
\title{Make filtering information information data frame from JSON file}
\usage{
read_length_hist(df, groupName, readLength, nRead, col)
}
\arguments{
\item{json}{A JSON file with QC information}
}
\value{
A data frame with FASTQ file information
}
\description{
Make filtering information information data frame from JSON file
}
\examples{
\dontrun{
# Input

# Analysis

# Output
}

}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
