% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{readLengthFastq}
\alias{readLengthFastq}
\title{Extract reads length information from a fastq file}
\usage{
readLengthFastq(fq)
}
\arguments{
\item{fq}{A fastq file}
}
\value{
A data frame with reads length (read_length)
and number of reads (num_reads)
}
\description{
Extract reads length information from a fastq file
}
\examples{
# Input
fq1 <- system.file("extdata", "Fox3_Std_small.fq.gz", package = "Rfastp")

# Analysis
df <- readLengthFastq(fq = fq1)

# Output
df
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
