% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{readLengthHist}
\alias{readLengthHist}
\title{Read length distribution (barplot)}
\usage{
readLengthHist(
  fq = NULL,
  df = NULL,
  read_length = "read_length",
  num_reads = "num_reads",
  col = "red",
  name = "Not trimmed"
)
}
\arguments{
\item{fq}{A fastq file}

\item{df}{A data frame with reads length and number of reads}

\item{read_length}{Column with reads length}

\item{num_reads}{Column with number of reads of a particulat length}

\item{col}{Color of the bars}

\item{name}{Either the fastq file is trimmed or not-trimmed}
}
\value{
A data frame with FASTQ file information
}
\description{
Read length distribution (barplot)
}
\examples{
# Input
fq1 <- system.file("extdata", "Fox3_Std_small.fq.gz", package = "Rfastp")

# Analysis
p <- readLengthHist(fq = fq1)

# Output
p
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
