% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{reads_duplication_plot}
\alias{reads_duplication_plot}
\title{Duplicated reads plot}
\usage{
reads_duplication_plot(df, dupRateCol, GCpercentCol, dupLevelCol)
}
\arguments{
\item{df}{A data frame}

\item{dupRateCol}{Column with duplication rate in df}

\item{GCpercentCol}{Column with GC percentage in df}

\item{dupLevelCol}{Column with duplication levels in df}
}
\value{
An interactive plot
}
\description{
Duplicated reads plot
}
\examples{
\dontrun{
# Input

# Analysis

# Output

}

}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
