% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{reads_quality_plot}
\alias{reads_quality_plot}
\title{Make filtering information information data frame from JSON file}
\usage{
reads_quality_plot(df, bpPosCol, A_col, T_col, G_col, C_col, meanQual = NULL)
}
\arguments{
\item{json}{A JSON file with QC information}
}
\value{
A data frame with FASTQ file information
}
\description{
Make filtering information information data frame from JSON file
}
\examples{
\dontrun{
# Input

# Analysis

# Output
}

}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
