% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{readsQualityPlot}
\alias{readsQualityPlot}
\title{Read quality score plot}
\usage{
readsQualityPlot(
  json = NULL,
  df = NULL,
  bpPos = "bpPos",
  Aa = "Aa",
  Tt = "Tt",
  Gg = "Gg",
  Cc = "Cc",
  meanQual = NULL
)
}
\arguments{
\item{json}{A JSON file with QC information from \code{\link{qcFastq}}}

\item{df}{A data frame}

\item{bpPos}{Column name with base-pair position}

\item{Aa}{Column with Adenine quality score}

\item{Tt}{Column with Thymine quality score}

\item{Gg}{Column with Guanine quality score}

\item{Cc}{Column with Cytosine quality score}

\item{meanQual}{Column with `mean` quality score}
}
\value{
An interactive plot
}
\description{
Read quality score plot
}
\examples{
# Input
fq1 <- system.file("extdata", "Fox3_Std_small.fq.gz", package = "Rfastp")
outDir <- tempdir()

# Analysis
qc_res <- qc_fastq(file = fq1, outDir = outDir)
p <- readsQualityPlot(json = qc_res$json)

# Output
p
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
