% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{removeCorrelatedFeatures}
\alias{removeCorrelatedFeatures}
\title{removeCorrelatedFeatures}
\usage{
removeCorrelatedFeatures(counts, o, cort = 0.6, cvt = 0.1, minmc = 5,
  minmdiff = 1)
}
\arguments{
\item{counts}{A count matrix from which to remove correlated subfeatures.}

\item{o}{An object of class smallRNAexp}

\item{cort}{Max feature correlation (default 0.6) for splitting}

\item{cvt}{Max coefficient of variation (default 0.1) for splitting}

\item{minmc}{Min mean count (default 5) for splitting}

\item{minmdiff}{Min mean difference (default 1) for splitting}
}
\value{
A filtered version of `counts`
}
\description{
Writes the results of a differential expression analysis on a smallRNAexp object in an excel file.
}
