% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{repSoftClip}
\alias{repSoftClip}
\title{Dealing with soft-clipping}
\usage{
repSoftClip(GA)
}
\arguments{
\item{GA}{`GAlignment file`}
}
\value{
`GRanges`
}
\description{
Dealing with soft-clipping
}
