% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{runDEA}
\alias{runDEA}
\title{runDEA}
\usage{
runDEA(
  o,
  formula = NULL,
  groups = NULL,
  method = "edgeR",
  types = NULL,
  normalizeByType = FALSE,
  replicates = NULL,
  forceGLM = FALSE,
  coef = NULL,
  filterFun = function(x) {     sum(x > 20) > 2 },
  ...
)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{formula}{The formula to test. Either this or `groups` should be given.}

\item{groups}{Either NULL (if `formula` is given), or the name of a column in phenoData, or a vector of length equal to the number of samples.}

\item{method}{DEA method, either "edgeR", "voom", or "DESeq2".}

\item{types}{The types of RNA to test (default all). A separate analysis will anyway be performed for miRNAs and piRNAs, but this parameter dictates the RNAs that are included in the global analyses.}

\item{normalizeByType}{Logical; whether each RNA type should be analyzed on its own (default FALSE). If TRUE, the default normalization method is used.}

\item{replicates}{An optional vector indicating (technical or 'incomplete') replicates (ignored unless method="voom")}

\item{forceGLM}{Whether to force using GLM even when doing a simple comparison between two groups.}

\item{coef}{Coefficient to be tested (defaults to the last one of the formula).}

\item{...}{Further arguments passed to the removeCorrelatedFeatures function.}
}
\value{
A list containing the DEA results on different sets of features.
}
\description{
Performs differential expression analysis (DEA) on a shortRNAexp object. This is a wrapper around common RNAseq DEA packages.
}
