% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortRNAexp.R
\name{shortRNAexp}
\alias{shortRNAexp}
\title{shortRNAexp}
\usage{
shortRNAexp(seqcounts, alignments, annotation, phenoData = NULL,
  description = NA_character_, rules = defaultAssignRules(),
  keepAllSrcs = TRUE)
}
\arguments{
\item{seqcounts}{Either a numeric matrix (with samples as columns and short sequences as rows) or the path to such a file.}

\item{alignments}{An alignment file (on the custom genome created along with the `annotation`) for each unique sequence, in BAM format.}

\item{annotation}{A feature annotation object of type `GRanges`, as created by the `prepareAnnotation` function.}

\item{phenoData}{Either a data.frame of phenotypic data (with samples as rows and features as columns) or the path to such a file.}

\item{description}{Character vector describing the experiment/dataset.}

\item{rules}{Assignment rules (see `?defaultAssignRules`).}

\item{keepAllSrcs}{Logical; whether to keep information on all sequences' potential sources (before assignment).}
}
\value{
a shortRNAexp object.
}
\description{
Creates a shortRNAexp object
}
