% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{shortRNAexp_align}
\alias{shortRNAexp_align}
\title{shortRNAexp_align}
\usage{
shortRNAexp_align(
  fasta,
  outputfile,
  bowtie1index,
  starindex,
  bowtie1 = "bowtie",
  star = "STAR",
  samtools = "samtools",
  m = 1000,
  nthreads = 4
)
}
\arguments{
\item{fasta}{Path to a fasta file containing the unique reads.}

\item{outputfile}{Path to the output bam file.}

\item{bowtie1index}{Path to the base of the bowtie1 index.}

\item{starindex}{Path to the STAR index folder.}

\item{bowtie1}{Executable for bowtie1 (default 'bowtie').}

\item{star}{Executable for STAR (default 'STAR').}

\item{samtools}{Executable for samtools (default 'samtools').}

\item{m}{Maximum alignments for a read to be considered (default 1000).}

\item{nthreads}{Number of threads for alignment (default 4).}
}
\description{
A wrapper for the 2-steps alignment method.
}
