% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallRNAexp.R
\name{smallRNAexp}
\alias{smallRNAexp}
\title{smallRNAexp}
\usage{
smallRNAexp(seqcounts, sources, phenoData, features2meta = NULL,
  description = NA_character_, rules = defaultAssignRules(),
  keepAllSrcs = TRUE)
}
\arguments{
\item{seqcounts}{Either a numeric matrix (with samples as columns and short sequences as rows) or the path to such a file.}

\item{sources}{Either a data.frame or sequence sources or the path to such a file.}

\item{phenoData}{Either a data.frame of phenotypic data (with samples as rows and features as columns) or the path to such a file.}

\item{features2meta}{Optional data.frame or file for establishing equivalences between features (should inclue two columns, in order: original_gene_name, replacement_name).}

\item{description}{Character vector describing the experiment/dataset.}

\item{rules}{Assignment rules (see `?defaultAssignRules`).}

\item{keepAllSrcs}{Logical; whether to keep information on all sequences' potential sources (before assignment).}
}
\value{
a smallRNAexp object.
}
\description{
Creates a smallRNAexp object
}
