% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{smallRNAexp_parseBam}
\alias{smallRNAexp_parseBam}
\title{smallRNAexp_parseBam}
\usage{
smallRNAexp_parseBam(bam, elements, outputfile = NULL, shell = "bash",
  samtools = "samtools", bedtools = "bedtools")
}
\arguments{
\item{bam}{Path to the sorted bam file.}

\item{elements}{Path to the genomic elements bed file.}

\item{outputfile}{Path for the output file.}

\item{shell}{Shell executable, default 'bash'.}

\item{samtools}{Executable for samtools (default 'samtools').}

\item{bedtools}{Executable for bedtools (default 'bedtools').}
}
\description{
Extracts reads' possible sources from an alignment file.
}
