% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeFL.R
\name{tRNAtoFL}
\alias{tRNAtoFL}
\title{Convert tRNAs annotation GRanges to a Factor List}
\usage{
tRNAtoFL(gr, sp = "mm10")
}
\arguments{
\item{gr}{annotation GRanges with columns: tx_id, tx_biotype, and symbol}

\item{sp}{name of the species}
}
\value{

}
\description{
Convert tRNAs annotation GRanges to a Factor List
}
