% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{tryAdapters}
\alias{tryAdapters}
\title{tryAdapters}
\usage{
tryAdapters(
  fq1,
  fq2 = NULL,
  adapters = NULL,
  addReverseComplement = TRUE,
  maxReads = 5e+05
)
}
\arguments{
\item{fq1}{Path to a fastq (or fastq.gz) file containing Read1.}

\item{fq2}{Path to a fastq (or fastq.gz) file containing Read 2 (for
paired-end data).}

\item{adapters}{Path to a fasta file containing adapter sequences. If
not provided, a set of common adapters will be used.}

\item{addReverseComplement}{Whether to also test the reverse-complement of
the adapters (default TRUE).}

\item{maxReads}{The maximum number of reads to use (default 500k)}
}
\value{
A list.
}
\description{
Looks for multiple 3'-end adapters in reads, and identifies the top ones.
}
