% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umic.R
\name{umiCollapse}
\alias{umiCollapse}
\title{umiCollapse
This function is adapted from the code available on: 
https://github.com/BiodataAnalysisGroup/UMIc. The function is adapted as the
original code was throwing some errors and was extremely slow for big `fastq`
files. Here, we try to make it a bit faster.}
\usage{
umiCollapse(
  fastq,
  UMIlocation = "R1",
  UMIlength = 8,
  sequenceLength = 51,
  countsCutoff = 5,
  UMIdistance = 1,
  sequenceDistance = 3,
  outDir = getwd()
)
}
\arguments{
\item{fastq}{path to a fastq file}

\item{UMIlocation}{UMI located. Default: in Read1 --> "R1"}

\item{UMIlength}{UMI length in base-pair. Default: 8}

\item{sequenceLength}{Length of sequences. Default: 51bp}

\item{countsCutoff}{Minimum read counts per UMI, for initial data cleaning.
Default: 5}

\item{UMIdistance}{Maximum UMI distance for UMI merging. Default: 1}

\item{sequenceDistance}{Maximum sequence distance for UMI correction.
Default: 3}

\item{outDir}{Output directory for results. Default: current working
directory}
}
\value{

}
\description{
umiCollapse
This function is adapted from the code available on: 
https://github.com/BiodataAnalysisGroup/UMIc. The function is adapted as the
original code was throwing some errors and was extremely slow for big `fastq`
files. Here, we try to make it a bit faster.
}
