% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{volcano}
\alias{volcano}
\title{volcano}
\usage{
volcano(res, writeTop = 0)
}
\arguments{
\item{res}{The results data.frame}

\item{writeTop}{Number of top features to write on the plot (default 0).}
}
\description{
Plots a volcano plot from a DEA result data.frame
}
