% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{writeDEA}
\alias{writeDEA}
\title{writeDEA}
\usage{
writeDEA(res, file = "DEA.xlsx", fdr.threshold = NULL)
}
\arguments{
\item{res}{A list such as produced by the `runDEA` function.}

\item{file}{Path to the file where to save the results. Defaults to DEA.xlsx in the working directory.}

\item{fdr.threshold}{Either NULL (default) or a numeric value between 0 and 1 indicating the maximum FDR value for a feature to be included in the file.}
}
\description{
Writes the results of a differential expression analysis on a smallRNAexp object in an excel file.
}
