% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqCounts.R
\name{writeUniqueFasta}
\alias{writeUniqueFasta}
\title{writeUniqueFasta}
\usage{
writeUniqueFasta(countMatrix, fasta_file = "unique_fragments.fasta")
}
\arguments{
\item{countMatrix}{A fragment count matrix, with sequences as row.names, as
produced by `fastq2SeqCountMatrix()`}

\item{fasta_file}{The filepath to which to save the fasta file.}
}
\description{
writeUniqueFasta
}
