% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{aggregateSequences}
\alias{aggregateSequences}
\title{aggregateSequences}
\usage{
aggregateSequences(o, quiet = FALSE)
}
\arguments{
\item{o}{An object of class smallRNAexp.}

\item{quiet}{Logical; indicates whether to suppresses messages (default FALSE).}
}
\value{
An updated object of class smallRNAexp.
}
\description{
Aggregates individual sequences into features
}
