% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsDKT.R
\name{alignShortRNA}
\alias{alignShortRNA}
\title{Read Bismark coverage files containing methylated and unmethylated read counts for CpG loci and create DGEList.}
\usage{
alignShortRNA(fastq, index)
}
\arguments{
\item{files}{character vector of file names.}

\item{nParallel}{numeric value of how many cores to be used for reading files. Important when reading files with data.table package.}

\item{verbose}{logical. If TRUE, read progress messages are send to standard output.}
}
\value{
A `DGEList`.
}
\description{
Read Bismark coverage files containing methylated and unmethylated read counts for CpG loci and create DGEList.
}
\examples{
dge.files <- ReadBismark2DGE(files = files, sample.names = names, data.table = T, nParallel = 8)
}
\seealso{
Rsubread
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
