% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignRead}
\alias{assignRead}
\title{assignRead}
\usage{
assignRead(sources, rules = defaultAssignRules())
}
\arguments{
\item{sources}{A data.frame of possible sources for the read (i.e., a subset of the `sources` slot of a shortRNAexp object).}

\item{rules}{A list of assignment rules; defaults to `defaultAssignRules()`.}
}
\value{
A vector including the following variables: seq, location, cigar, pos_in_feature, feature_length, src name, src type, overlap length, src length, assignment status
}
\description{
Assigns read to one or more of possible sources.
}
