% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignReads}
\alias{assignReads}
\title{assignReads}
\usage{
assignReads(sources, rules = defaultAssignRules())
}
\arguments{
\item{sources}{A data.frame of overlaps, as produced by
\code{\link{overlapWithTx}}}

\item{rules}{Assignment rules (see \code{\link{defaultAssignmentRules}})}
}
\value{
A \code{\link[S4Vectors]{DataFrame-class}}
}
\description{
assignReads
}
