% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignReads}
\alias{assignReads}
\title{assignReads}
\usage{

assignReads(sources, rules = defaultAssignRules(), trees = list(),
  nthreads = NULL)
}
\arguments{
\item{sources}{A data.frame of possible sources, as produced by the findOverlap function.}

\item{rules}{A list of assignment rules; defaults to `defaultAssignRules()`.}

\item{nthreads}{The number of threads to use; default to detected cores -1 if the parallel package is installed.}
}
\value{
A data.frame of assigned sources.
}
\description{
Assigns each read to one or more of possible sources.
}
