% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{autoLayout}
\alias{autoLayout}
\title{autoLayout}
\usage{
autoLayout(nb, byrow = FALSE)
}
\arguments{
\item{nb}{Minimum number of frames}

\item{byrow}{Whether to fill frames by row instead of by column
(default FALSE, i.e. fill by column)}
}
\value{
The total number of frames in the layout.
}
\description{
Creates a layout for a given (minimum) number of frames
}
