% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{checkPiRNA}
\alias{checkPiRNA}
\title{checkPiRNA}
\usage{
checkPiRNA(seqs, type = "primary", size = 25:35)
}
\arguments{
\item{seqs}{A character vector of sequences}
}
\value{
A logical vector
}
\description{
Checks whether a sequence could in principle be a piRNA
}
