% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{getOverlapValidity}
\alias{getOverlapValidity}
\title{getOverlapValidity}
\usage{
getOverlapValidity(sources, rules = defaultAssignRules(), tryFallback = TRUE)
}
\arguments{
\item{sources}{the output of overlapWithTx2}

\item{rules}{a named list of assignment rules}

\item{tryFallback}{Logical; whether to test the fallback type.}
}
\value{
Data table (same format as `sources`)
}
\description{
getOverlapValidity
}
