% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{getPotentialSources}
\alias{getPotentialSources}
\title{getPotentialSources}
\usage{
getPotentialSources(o, seq)
}
\arguments{
\item{o}{An object of class shortRNAexp.}

\item{seq}{A character vector including the sequence(s) of interest.}
}
\value{
A data.frame.
}
\description{
Returns the potential sources (before assignment) of a sequence of interest.
}
