% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortRNAexp.R
\name{getSeqCounts}
\alias{getSeqCounts}
\title{getSeqCounts}
\usage{
getSeqCounts(o, type = NULL, status = c("unknown", "invalid",
  "ambiguous", "unique"), normalized = FALSE, formatCase = FALSE)
}
\arguments{
\item{o}{An object of class shortRNAexp}

\item{type}{Types of sequences to be fetched (e.g. miRNA, tRNA, etc.), fetches all by default.}

\item{status}{Character vector of seq statuses to be selected (any combination of "unmapped","invalid", "unknown", "ambiguous", and "unique"). Defaults to all mapped sequences.}

\item{normalized}{Logical; whether to return the normalized counts (default FALSE).}

\item{formatCase}{Logical; whether to format sequences' case on the basis of clipping (default FALSE).}
}
\value{
A count matrix.
}
\description{
Returns a count matrix for requested sequences
}
