% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_plg.R
\name{longestOrderedOverlap}
\alias{longestOrderedOverlap}
\title{longestOrderedOverlap}
\usage{
longestOrderedOverlap(x)
}
\arguments{
\item{x}{An AtomicList}
}
\value{
An atomic vector
}
\description{
Returns the longest ordered overlap between elements of an atomic list
}
