% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{make_dup_plot}
\alias{make_dup_plot}
\title{Make duplicated reads plot from JSON file}
\usage{
make_dup_plot(json)
}
\arguments{
\item{json}{A JSON file}
}
\value{
An interactive plot

#' @examples
\dontrun{
# Input

# Analysis

# Output

}
}
\description{
Make duplicated reads plot from JSON file
}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
