% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoprep.R
\name{prepareAnnotation}
\alias{prepareAnnotation}
\title{prepareAnnotation}
\usage{
prepareAnnotation(species = NULL, filelist = NULL, destination = getwd(),
  description = "")
}
\arguments{
\item{species}{The species for automatic fetching and creation of the annotation. See `?getAnnotationFiles` for the supported species/assemblies.}

\item{filelist}{A list of named feature files, as produced by 'getAnnotationFiles', indicating the necessary files. The following slots are required: 'features.gtf','features.gtf','mirbase.gtf','gtRNAdb.fa','gtRNAdb.bed','tRNA.gtf'; the 'pi_precursors' is optional.}

\item{destination}{The destination folder for the files (defaults to current working directory)}

\item{description}{An optional character vector containing a description of the annotation (default empty)}
}
\value{
A GRanges object, and produces modified fasta files.
}
\description{
Prepares an annotation GRanges and modified miRNA/tRNA sequences for a given organism. Either `filelist` or `species` should be provided.
}
