% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{reads_duplication_plot}
\alias{reads_duplication_plot}
\title{Duplicated reads plot}
\usage{
reads_duplication_plot(df, dupRateCol, GCpercentCol = NULL, dupLevelCol)
}
\arguments{
\item{df}{A data frame}

\item{dupRateCol}{Column with duplication rate in df}

\item{GCpercentCol}{Column with GC percentage in df}

\item{dupLevelCol}{Column with duplication levels in df}
}
\value{
An interactive plot
}
\description{
Duplicated reads plot
}
\examples{
# Input
df <- data.frame(
  dupRate = rev(sort(c(70, 10, 1:5, rep(0, 13)))),
  dupLevel = 1:20
)
df$GCpercent <- c(runif(n = 5, min = 51, max = 55), rep(0, 15))

# Analysis
p <- reads_duplication_plot(
  df = df,
  dupRateCol = "dupRate",
  GCpercentCol = NULL,
  dupLevelCol = "dupLevel"
)

p_GC <- reads_duplication_plot(
  df = df,
  dupRateCol = "dupRate",
  GCpercentCol = "GCpercent",
  dupLevelCol = "dupLevel"
)

# Output
p
p_GC
}
\seealso{
Rfastp
}
\author{
Deepak Tanwar (tanward@ethz.ch)
}
