% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{tRFtype}
\alias{tRFtype}
\title{tRFtype}
\usage{
tRFtype(srcs, rules = defaultAssignRules())
}
\arguments{
\item{o}{An object of class shortRNAexp.}

\item{seq}{A character vector including the sequence(s) of interest.}
}
\value{
A data.frame.
}
\description{
Returns the tRNA fragment type from a sources data.frame
}
